xed: Add support for reading symbol tables from nm style dumps
When decoding Linux kernel dumps it can be useful to use the
symbol table in /proc/kallsyms. This is the same format
as output by `nm'. Add a new -S option to xed that reads
such a file and uses it to resolve symbols.
Right now this is only implemented in filter mode (-F)