Perf filter 4 (#37)
* Add support for disassembling as a filter

Linux perf can output hex dumps of individual instructions, but it
doesn't have a full disassembler. Extend xed to allow filtering of
such dumps.

The instructions always have a specific format:

other file HEX-IP(optional) ... insn: byte1 byte2 ... other things

This patch adds a new -F prefix mode that looks for this
format and runs as a filter that replaces the insn: ... patterns
with disassembled instructions.

* xed: Add support for reading symbol tables from nm style dumps

When decoding Linux kernel dumps it can be useful to use the
symbol table in /proc/kallsyms. This is the same format
as output by `nm'. Add a new -S option to xed that reads
such a file and uses it to resolve symbols.
Right now this is only implemented in filter mode (-F)