Add -q command-line knob option to asmparse
When -q is active, only the resulting bytes are printed to stdout, and no
diagnostics of any kind. This simplified consumption of the output by external
tools.

If an error happened, problem details are printed to stderr, and error code
is non-zero, again to simplify integration with external tools.

Internally, verbosity levels 0, 1 and 2 are used: default is 1, -q is 0, -v is 2