Remove VERSION file
XED versioning format is changed to a date base format for external releases.
For the internal repo, XED version information should be taken only from git tags.
The VERSION file is only exists in the external repository, and it's auto generated
by the external release automation job.

In order to support the above, we check the VERSION file only after checking git tags.
If both methods fails, we use a default non-version string of "000".